Do a major rewrite of the application logic
Initially, there was an ExpensesTable component that was a parent to
ExpensesForm. Using this approach, changes made in ExpensesForm were not
reflected automatically in ExpensesTable (as data flows from parent to
child in React, not the other way round). Thus, I had to merge both
components into ExpensesForm, and delete a number of unused components
that are children of ExpensesTable